refactor: move `ArtworkTypeIDs` to `lib/MB/CoverArt`
It'll most likely be useful in Supercharged CAA Edits, but more
importantly, we'll need it in some of the E2E tests, and importing
it from `providers/base` (which wasn't a good spot for the enum
anyway) leads to errors because of a dependency chain ending up
in `@src/compat`, which dereferences `GM` functions.
