+++ src/lib/MB/CoverArt.ts
@@ -1,0 +1,16 @@
+export enum ArtworkTypeIDs {
+    Back = 2,
+    Booklet = 3,
+    Front = 1,
+    Liner = 12,
+    Medium = 4,
+    Obi = 5,
+    Other = 8,
+    Poster = 11,
+    Raw = 14,  // Raw/Unedited
+    Spine = 6,
+    Sticker = 10,
+    Track = 7,
+    Tray = 9,
+    Watermark = 13
+}
