+++ src/mb_enhanced_cover_art_uploads/providers/base.ts
@@ -1,4 +1,5 @@
 import { LOGGER } from '@lib/logging/logger';
+import { ArtworkTypeIDs } from '@lib/MB/CoverArt';
 import { filterNonNull, groupBy } from '@lib/util/array';
 import { blobToDigest } from '@lib/util/blob';
 import { parseDOM, qs } from '@lib/util/dom';
@@ -131,23 +132,6 @@ export interface CoverArt {
 export interface ParsedTrackImage {
     url: string;
     trackNumber?: string;
-}
-
-export enum ArtworkTypeIDs {
-    Back = 2,
-    Booklet = 3,
-    Front = 1,
-    Liner = 12,
-    Medium = 4,
-    Obi = 5,
-    Other = 8,
-    Poster = 11,
-    Raw = 14,  // Raw/Unedited
-    Spine = 6,
-    Sticker = 10,
-    Track = 7,
-    Tray = 9,
-    Watermark = 13,
 }
 
 export abstract class HeadMetaPropertyProvider extends CoverArtProvider {
