+++ tests/unit/mb_enhanced_cover_art_uploads/providers/7digital.test.ts
@@ -1,4 +1,4 @@
-import { ArtworkTypeIDs } from '@src/mb_enhanced_cover_art_uploads/providers/base';
+import { ArtworkTypeIDs } from '@lib/MB/CoverArt';
 import { SevenDigitalProvider } from '@src/mb_enhanced_cover_art_uploads/providers/7digital';
 
 import { itBehavesLike } from '@test-utils/shared_behaviour';
