+++ tests/unit/mb_enhanced_cover_art_uploads/providers/rateyourmusic.test.ts
@@ -1,5 +1,5 @@
+import { ArtworkTypeIDs } from '@lib/MB/CoverArt';
 import type { CoverArtProvider } from '@src/mb_enhanced_cover_art_uploads/providers/base';
-import { ArtworkTypeIDs } from '@src/mb_enhanced_cover_art_uploads/providers/base';
 import { RateYourMusicProvider } from '@src/mb_enhanced_cover_art_uploads/providers/rateyourmusic';
 
 import { setupPolly } from '@test-utils/pollyjs';
