+++ tests/unit/mb_enhanced_cover_art_uploads/providers/soundcloud.test.ts
@@ -1,4 +1,4 @@
-import { ArtworkTypeIDs } from '@src/mb_enhanced_cover_art_uploads/providers/base';
+import { ArtworkTypeIDs } from '@lib/MB/CoverArt';
 import { SoundcloudProvider } from '@src/mb_enhanced_cover_art_uploads/providers/soundcloud';
 
 import { itBehavesLike } from '@test-utils/shared_behaviour';
