+++ tests/unit/mb_enhanced_cover_art_uploads/providers/spotify.test.ts
@@ -1,4 +1,4 @@
-import { ArtworkTypeIDs } from '@src/mb_enhanced_cover_art_uploads/providers/base';
+import { ArtworkTypeIDs } from '@lib/MB/CoverArt';
 import { SpotifyProvider } from '@src/mb_enhanced_cover_art_uploads/providers/spotify';
 
 import { itBehavesLike } from '@test-utils/shared_behaviour';
