+++ tests/unit/mb_enhanced_cover_art_uploads/seeding/parameters.test.ts
@@ -1,4 +1,4 @@
-import { ArtworkTypeIDs } from '@src/mb_enhanced_cover_art_uploads/providers/base';
+import { ArtworkTypeIDs } from '@lib/MB/CoverArt';
 import { SeedParameters } from '@src/mb_enhanced_cover_art_uploads/seeding/parameters';
 
 describe('seed parameters', () => {
